chore: move subscription to user schema
