+++ template/apps/api/src/resources/subscription/actions/cancel.ts
@@ -1,22 +1,18 @@
 import stripe from 'services/stripe/stripe.service';
 
-import { subscriptionService } from 'resources/subscription';
-
 import { AppKoaContext, AppRouter } from 'types';
 
 async function handler(ctx: AppKoaContext) {
   const { user } = ctx.state;
 
-  const currentSubscription = await subscriptionService.findOne({ customer: user.stripeId || undefined });
-
-  if (!currentSubscription) {
+  if (!user.subscription) {
     ctx.status = 400;
     ctx.message = 'Subscription does not exist';
 
     return;
   }
 
-  await stripe.subscriptions.update(currentSubscription.subscriptionId, {
+  await stripe.subscriptions.update(user.subscription.subscriptionId, {
     cancel_at_period_end: true,
   });
