+++ template/apps/api/src/resources/subscription/actions/preview-upgrade.ts
@@ -1,8 +1,6 @@
 import { z } from 'zod';
 import stripe from 'services/stripe/stripe.service';
 
-import { subscriptionService } from 'resources/subscription';
-
 import { validateMiddleware } from 'middlewares';
 import { AppKoaContext, AppRouter } from 'types';
 
@@ -18,16 +16,14 @@ async function handler(ctx: AppKoaContext<ValidatedData>) {
   const { user } = ctx.state;
   const { priceId } = ctx.validatedData;
 
-  const currentSubscription = await subscriptionService.findOne({ customer: user.stripeId || undefined });
-
-  if (!currentSubscription) {
+  if (!user.subscription) {
     ctx.status = 400;
     ctx.message = 'Subscription does not exist';
 
     return;
   }
 
-  const subscriptionDetails = await stripe.subscriptions.retrieve(currentSubscription.subscriptionId);
+  const subscriptionDetails = await stripe.subscriptions.retrieve(user.subscription.subscriptionId);
 
   let items: any;
 
@@ -36,7 +32,7 @@ async function handler(ctx: AppKoaContext<ValidatedData>) {
       id: subscriptionDetails.items.data[0].id,
       price_data: {
         currency: 'USD',
-        product: currentSubscription.productId,
+        product: user.subscription.productId,
         recurring: {
           interval: subscriptionDetails.items.data[0].price.recurring?.interval,
           interval_count: 1,
@@ -53,7 +49,7 @@ async function handler(ctx: AppKoaContext<ValidatedData>) {
 
   const invoice = await stripe.invoices.retrieveUpcoming({
     customer: user.stripeId || undefined,
-    subscription: currentSubscription.subscriptionId,
+    subscription: user.subscription.subscriptionId,
     subscription_items: items,
     subscription_proration_behavior: 'always_invoice',
   });
