+++ template/apps/api/src/resources/subscription/index.ts
@@ -1,9 +1,7 @@
-import subscriptionService from './subscription.service';
 import subscriptionRoutes from './subscription.routes';
 
 export * from './subscription.types';
 
 export {
-  subscriptionService,
   subscriptionRoutes,
 };
