+++ template/apps/api/src/resources/user/user.schema.ts
@@ -1,5 +1,7 @@
 import { z } from 'zod';
 
+import subscriptionSchema from 'resources/subscription/subscription.schema';
+
 const schema = z.object({
   _id: z.string(),
   createdOn: z.date().optional(),
@@ -19,6 +21,7 @@ const schema = z.object({
   oauth: z.object({
     google: z.boolean().default(false),
   }).optional(),
+  subscription: subscriptionSchema.optional().nullable(),
 }).strict();
 
 export default schema;
