+++ template/apps/api/src/resources/user/user.service.ts
@@ -36,10 +36,48 @@ const attachStripeCustomerId = (data: any) => {
   );
 };
 
+const updateSubscription = async (data: any) => {
+  const subscription = {
+    subscriptionId: data.id,
+    priceId: data.plan.id,
+    productId: data.plan.product,
+    status: data.status,
+    interval: data.plan.interval,
+    currentPeriodStartDate: data.current_period_start,
+    currentPeriodEndDate: data.current_period_end,
+    cancelAtPeriodEnd: data.cancel_at_period_end,
+  };
+
+  service.atomic.updateOne(
+    {
+      stripeId: data.customer,
+    },
+    {
+      $set: {
+        subscription,
+      },
+    },
+  );
+};
+
+const deleteSubscription = async (data: any) => {
+  service.atomic.updateOne(
+    {
+      stripeId: data.customer,
+    },
+    {
+      $unset: {
+        subscription: '',
+      },
+    },
+  );
+};
+
 const privateFields = [
   'passwordHash',
   'signupToken',
   'resetPasswordToken',
+  'subscription',
 ];
 
 const getPublic = (user: User | null) => _.omit(user, privateFields);
@@ -48,4 +86,6 @@ export default Object.assign(service, {
   updateLastRequest,
   getPublic,
   attachStripeCustomerId,
+  updateSubscription,
+  deleteSubscription,
 });
